cloud-run-worker/requirements.txt
Flask==3.0.3
gunicorn==21.2.0
google-generativeai==0.7.2
youtube-transcript-api==0.6.2
yt-dlp==2024.08.06
tenacity==8.5.0
beautifulsoup4==4.12.3
